[ABS-551] Various fixes: Set the advance widths for all the nuqtas and diacritics to zero. Graphite is getting confused by them!
absBehMed.benn still needs "exit_behPreNoon" AP. I set it to (0,1200) and it was reasonable. 

Position of _hamza relative to absBehMed.benn is not right. When I changed the AP on the hamza to something like (0, -500) it was pretty good. 

absSadIni.qf needs a "markAbove" AP to attach the pesh. 

_dot3l needs a "markBelow" AP to attach the zair. 

Connection between absHehGoalMed and absAlefFin (exit_laamKaf) needs a bit of attention. 

Could you make the space quite a bit narrower? In Nafees it is more like 200m.